+++ CHANGELOG.md
@@ -3,6 +3,31 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## [1.1.0] - 2022-01-03
+
+### Added
+
+- The default JSON:API resource class can now be changed via
+  the `LaravelJsonApi\Laravel\LaravelJsonApi::defaultResource()` method. This should be set in a service
+  provider's `register()` method.
+- [#127](https://github.com/laravel-json-api/laravel/issues/127) The `JsonApiResource` class now has a
+  protected `serializeRelation` method that can be used to override the default serialization of relationships if
+  needed.
+- [#111](https://github.com/laravel-json-api/laravel/issues/111) Relationship documents returned by relationship `self`
+  routes will now include any non-standard links set on the resource relationship in the top-level `links` member.
+
+### Fixed
+
+- [#147](https://github.com/laravel-json-api/laravel/issues/147) Related relationship response now correctly merge the
+  relationship links into the top-level document links member.
+- [#130](https://github.com/laravel-json-api/laravel/issues/130) The `JsonApiResource` now correctly handles conditional
+  fields when iterating over relationships to find a specific relation.
+- [#105](https://github.com/laravel-json-api/laravel/issues/105) The JSON:API document returned by a relationship `self`
+  route now handles a relationship not existing if it is hidden. Previously an exception was thrown when attempting to
+  merge relationship links into the document.
+- [#111](https://github.com/laravel-json-api/laravel/issues/111) Relationship documents now handle a relationship that
+  does not have one or both of the `self` and `related` relationship links.
+
 ## [1.0.1] - 2021-12-08
 
 ### Changed
@@ -13,8 +38,8 @@ All notable changes to this project will be documented in this file. This projec
 ### Fixed
 
 - [#139](https://github.com/laravel-json-api/laravel/issues/139) Fix the `WhereHas` and `WhereDoesntHave` filters.
-  Previously these were not iterating over the filters from the correct resource schema - they were iterating over
-  the filters from the schema to which the relationship belonged. They now correctly iterate over the filters from the
+  Previously these were not iterating over the filters from the correct resource schema - they were iterating over the
+  filters from the schema to which the relationship belonged. They now correctly iterate over the filters from the
   schema for the resource that is on the inverse side of the relationship.
 
 ## [1.0.0] - 2021-07-31